Ask before overwriting already-existing problem with template

+++ template.py
@@ -1,25 +1,36 @@
 import os, sys, subprocess
 from distutils.dir_util import copy_tree
 
+def copyTreeCheck(source, destination):
+	if os.path.exists(destination):
+		x = input("The directory of this problem already exists, overwrite it? [y/n] ")
+		if x not in ["y", "Y"]:
+			return False
+
+	copy_tree(source, destination)
+	return True
+
 def create(source, problem):
 	destinationDir = None
 	if source in ["o", "olinfo"]:
 		destinationDir = f"./olinfo/{problem}/"
-		copy_tree("./src/template-olinfo/", destinationDir)
+		written = copyTreeCheck("./src/template-olinfo/", destinationDir)
 
-		readme = f"{destinationDir}README.md"
-		with open(readme, "r") as f: readmeCont = f.read()
-		readmeCont = "".join(readmeCont.split("<!--")[1:])
-		readmeCont = readmeCont.replace("$PROBLEM", f"{problem}")
-		with open(readme, "w") as f: f.write(readmeCont)
+		if written:
+			readme = f"{destinationDir}README.md"
+			with open(readme, "r") as f: readmeCont = f.read()
+			readmeCont = "".join(readmeCont.split("<!--")[1:])
+			readmeCont = readmeCont.replace("$PROBLEM", f"{problem}")
+			with open(readme, "w") as f: f.write(readmeCont)
 	elif source in ["u", "uva"]:
 		destinationDir = f"./uva/{problem}/"
-		copy_tree("./src/template-uva/", destinationDir)
+		written = copyTreeCheck("./src/template-uva/", destinationDir)
 
-		tasks = f"{destinationDir}.vscode/tasks.json"
-		with open(tasks, "r") as f: tasksCont = f.read()
-		tasksCont = tasksCont.replace("$PROBLEM", problem)
-		with open(tasks, "w") as f: f.write(tasksCont)
+		if written:
+			tasks = f"{destinationDir}.vscode/tasks.json"
+			with open(tasks, "r") as f: tasksCont = f.read()
+			tasksCont = tasksCont.replace("$PROBLEM", problem)
+			with open(tasks, "w") as f: f.write(tasksCont)
 	else:
 		print(f"Invalid source: {source}")
 		return
